Move defer() function into dedicated file
