+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(core_sources
     base/array_view.hpp
     base/container_utils.hpp
+    base/defer.hpp
     base/grid.hpp
     base/math_tools.hpp
     base/spatial_types.hpp
